Create solv.py
In this project, I worked with my team, we discussed to divide the tasks according to each person's strengths. My role in this project is to collect images from the camera, pre-process to give the most characteristic feature of the shape image, the bounding box of each color square on the rubik, and determine the color of the tiles there. In addition, I was also assigned to research on optical flow to track the direction of movement when rotating the rubik, but in this case it did not work really well.